requirements.txt
msal
python-dotenv